+++ python/cudf/cudf/tests/test_dataframe.py
@@ -3278,6 +3278,34 @@ def test_empty_dataframe_describe():
 def test_as_column_types():
     from cudf.dataframe import columnops
 
+    col = columnops.as_column(Series([]))
+    assert_eq(col.dtype, np.dtype("float64"))
+    gds = Series(col)
+    pds = pd.Series(pd.Series([]))
+
+    assert_eq(pds, gds)
+
+    col = columnops.as_column(Series([]), dtype="float32")
+    assert_eq(col.dtype, np.dtype("float32"))
+    gds = Series(col)
+    pds = pd.Series(pd.Series([], dtype="float32"))
+
+    assert_eq(pds, gds)
+
+    col = columnops.as_column(Series([]), dtype="str")
+    assert_eq(col.dtype, np.dtype("object"))
+    gds = Series(col)
+    pds = pd.Series(pd.Series([], dtype="str"))
+
+    assert_eq(pds, gds)
+
+    col = columnops.as_column(Series([]), dtype="object")
+    assert_eq(col.dtype, np.dtype("object"))
+    gds = Series(col)
+    pds = pd.Series(pd.Series([], dtype="object"))
+
+    assert_eq(pds, gds)
+
     pds = pd.Series(np.array([1, 2, 3]), dtype="float32")
     gds = Series(columnops.as_column(np.array([1, 2, 3]), dtype="float32"))
 
@@ -3285,6 +3313,32 @@ def test_as_column_types():
 
     pds = pd.Series([1, 2, 3], dtype="float32")
     gds = Series([1, 2, 3], dtype="float32")
+
+    assert_eq(pds, gds)
+
+    pds = pd.Series([])
+    gds = Series(columnops.as_column(pds))
+    assert_eq(pds, gds)
+
+    pds = pd.Series([1, 2, 4], dtype="int64")
+    gds = Series(columnops.as_column(Series([1, 2, 4]), dtype="int64"))
+
+    assert_eq(pds, gds)
+
+    pds = pd.Series([1.2, 18.0, 9.0], dtype="float32")
+    gds = Series(
+        columnops.as_column(Series([1.2, 18.0, 9.0]), dtype="float32")
+    )
+
+    assert_eq(pds, gds)
+
+    pds = pd.Series([1.2, 18.0, 9.0], dtype="str")
+    gds = Series(columnops.as_column(Series([1.2, 18.0, 9.0]), dtype="str"))
+
+    assert_eq(pds, gds)
+
+    pds = pd.Series(pd.Index(["1", "18", "9"]), dtype="int")
+    gds = Series(gd.dataframe.index.StringIndex(["1", "18", "9"]), dtype="int")
 
     assert_eq(pds, gds)